+++ src/core/langs/cpp.ts
@@ -15,17 +15,13 @@
 // You should have received a copy of the GNU General Public License
 // along with cph-ng.  If not, see <https://www.gnu.org/licenses/>.
 
-import { access, constants, readFile } from 'fs/promises';
 import { type } from 'os';
 import { basename, extname, join } from 'path';
-import { l10n } from 'vscode';
-import Io from '../../helpers/io';
 import Logger from '../../helpers/logger';
-import ProcessExecutor, { AbortReason } from '../../helpers/processExecutor';
 import Settings from '../../modules/settings';
 import { extensionPath } from '../../utils/global';
-import { FileWithHash } from '../../utils/types';
-import { TCVerdicts } from '../../utils/types.backend';
+import { KnownResult, UnknownResult } from '../../utils/result';
+import { FileWithHash } from '../../utils/types.backend';
 import {
     CompileAdditionalData,
     DefaultCompileAdditionalData,
@@ -72,161 +68,78 @@ export class LangCpp extends Lang {
             forceCompile,
         );
         if (skip) {
-            return {
-                verdict: TCVerdicts.UKE,
-                data: { outputPath, hash },
-            };
+            return new UnknownResult({ outputPath, hash });
         }
 
-        const { objcopy, useWrapper, useHook, timeout } = Settings.compilation;
-        try {
-            const compileCommands: string[][] = [];
-            const postCommands: string[][] = [];
-            if (canUseWrapper && useWrapper) {
-                const obj = `${outputPath}.o`;
-                const wrapperObj = `${outputPath}.wrapper.o`;
-                const linkObjects = [obj, wrapperObj];
+        const { objcopy, useWrapper, useHook } = Settings.compilation;
+        const compileCommands: string[][] = [];
+        const postCommands: string[][] = [];
+        const compilerArgs = args.split(/\s+/).filter(Boolean);
+        if (canUseWrapper && useWrapper) {
+            const obj = `${outputPath}.o`;
+            const wrapperObj = `${outputPath}.wrapper.o`;
+            const linkObjects = [obj, wrapperObj];
 
-                const compilerArgs = args.split(/\s+/).filter(Boolean);
-                compileCommands.push(
-                    [compiler, src.path, ...compilerArgs, '-c', '-o', obj],
-                    [
-                        compiler,
-                        '-fPIC',
-                        '-c',
-                        join(extensionPath, 'res', 'wrapper.cpp'),
-                        '-o',
-                        wrapperObj,
-                    ],
-                );
-                if (useHook) {
-                    const hookObj = `${outputPath}.hook.o`;
-                    linkObjects.push(hookObj);
-                    compileCommands.push([
-                        compiler,
-                        '-fPIC',
-                        '-Wno-attributes',
-                        '-c',
-                        join(extensionPath, 'res', 'hook.cpp'),
-                        '-o',
-                        hookObj,
-                    ]);
-                }
-                postCommands.push(
-                    [objcopy, '--redefine-sym', 'main=original_main', obj],
-                    [
-                        compiler,
-                        ...linkObjects,
-                        ...compilerArgs,
-                        '-o',
-                        outputPath,
-                        ...(type() === 'Linux' ? ['-ldl'] : []),
-                    ],
-                );
-            } else {
-                const compilerArgs = args.split(/\s+/).filter(Boolean);
-                const cmd = [
+            compileCommands.push(
+                [compiler, src.path, ...compilerArgs, '-c', '-o', obj],
+                [
                     compiler,
-                    src.path,
+                    '-fPIC',
+                    '-c',
+                    join(extensionPath, 'res', 'wrapper.cpp'),
+                    '-o',
+                    wrapperObj,
+                ],
+            );
+            if (useHook) {
+                const hookObj = `${outputPath}.hook.o`;
+                linkObjects.push(hookObj);
+                compileCommands.push([
+                    compiler,
+                    '-fPIC',
+                    '-Wno-attributes',
+                    '-c',
+                    join(extensionPath, 'res', 'hook.cpp'),
+                    '-o',
+                    hookObj,
+                ]);
+            }
+            postCommands.push(
+                [objcopy, '--redefine-sym', 'main=original_main', obj],
+                [
+                    compiler,
+                    ...linkObjects,
                     ...compilerArgs,
                     '-o',
                     outputPath,
-                ];
-                if (Settings.runner.unlimitedStack && type() === 'Windows_NT') {
-                    cmd.push('-Wl,--stack,268435456');
-                }
-                debug && cmd.push('-g', '-O0');
-                compileCommands.push(cmd);
-            }
-
-            const compileResults = await Promise.all(
-                compileCommands.map((cmd) =>
-                    ProcessExecutor.execute({ cmd, ac, timeout }),
-                ),
+                    ...(type() === 'Linux' ? ['-ldl'] : []),
+                ],
             );
-            const results: string[] = [];
-            for (const result of compileResults) {
-                if (result instanceof Error) {
-                    return { verdict: TCVerdicts.SE, msg: result.message };
-                } else if (result.abortReason === AbortReason.UserAbort) {
-                    return {
-                        verdict: TCVerdicts.RJ,
-                        msg: l10n.t('Compilation aborted by user'),
-                    };
-                } else if (result.abortReason === AbortReason.Timeout) {
-                    return {
-                        verdict: TCVerdicts.CE,
-                        msg: l10n.t('Compilation timed out'),
-                    };
-                } else if (result.codeOrSignal) {
-                    Io.compilationMsg = await readFile(
-                        result.stderrPath,
-                        'utf-8',
-                    );
-                    return {
-                        verdict: TCVerdicts.CE,
-                        msg: l10n.t('Compiler exited with code {code}', {
-                            code: result.codeOrSignal,
-                        }),
-                    };
-                }
-                results.push(await readFile(result.stderrPath, 'utf-8'));
+        } else {
+            const cmd = [compiler, src.path, ...compilerArgs, '-o', outputPath];
+            if (Settings.runner.unlimitedStack && type() === 'Windows_NT') {
+                cmd.push('-Wl,--stack,268435456');
             }
-
-            for (const cmd of postCommands) {
-                const result = await ProcessExecutor.execute({
-                    cmd,
-                    ac,
-                    timeout,
-                });
-                if (result instanceof Error) {
-                    return { verdict: TCVerdicts.SE, msg: result.message };
-                } else if (result.abortReason === AbortReason.UserAbort) {
-                    return {
-                        verdict: TCVerdicts.RJ,
-                        msg: l10n.t('Compilation aborted by user'),
-                    };
-                } else if (result.abortReason === AbortReason.Timeout) {
-                    return {
-                        verdict: TCVerdicts.CE,
-                        msg: l10n.t('Compilation timed out'),
-                    };
-                } else if (result.codeOrSignal) {
-                    Io.compilationMsg = await readFile(
-                        result.stderrPath,
-                        'utf-8',
-                    );
-                    return {
-                        verdict: TCVerdicts.CE,
-                        msg: l10n.t('Compiler exited with code {code}', {
-                            code: result.codeOrSignal,
-                        }),
-                    };
-                }
-                results.push(await readFile(result.stderrPath, 'utf-8'));
-            }
-
-            this.logger.debug('Compilation completed successfully', {
-                path: src.path,
-                outputPath,
-            });
-            Io.compilationMsg = results
-                .map((result) => result.trim())
-                .filter((msg) => msg)
-                .join('\n\n');
-            return {
-                verdict: await access(outputPath, constants.X_OK)
-                    .then(() => TCVerdicts.UKE)
-                    .catch(() => TCVerdicts.CE),
-                data: { outputPath, hash },
-            };
-        } catch (e) {
-            this.logger.error('Compilation failed', e);
-            Io.compilationMsg = (e as Error).message;
-            return { verdict: TCVerdicts.CE, msg: '' };
+            debug && cmd.push('-g', '-O0');
+            compileCommands.push(cmd);
         }
+
+        for (const result of await Promise.all(
+            compileCommands.map((cmd) => this._executeCompiler(cmd, ac)),
+        )) {
+            if (result instanceof KnownResult) {
+                return result;
+            }
+        }
+        for (const cmd of postCommands) {
+            const result = await this._executeCompiler(cmd, ac);
+            if (result instanceof KnownResult) {
+                return result;
+            }
+        }
+        return new UnknownResult({ outputPath, hash });
     }
-    public async runCommand(target: string): Promise<string[]> {
+    public async getRunCommand(target: string): Promise<string[]> {
         this.logger.trace('runCommand', { target });
         return [target];
     }