+++ src/core/langs/lang.ts
@@ -17,30 +17,33 @@
 
 import { SHA256 } from 'crypto-js';
 import { readFile, unlink } from 'fs/promises';
-import { window } from 'vscode';
+import { l10n, window } from 'vscode';
+import { CompilationIo } from '../../helpers/io';
 import Logger from '../../helpers/logger';
-import ProcessExecutor from '../../helpers/processExecutor';
+import ProcessExecutor, { AbortReason } from '../../helpers/processExecutor';
+import Settings from '../../modules/settings';
 import { waitUntil } from '../../utils/global';
 import { exists } from '../../utils/process';
-import Result from '../../utils/result';
-import { FileWithHash } from '../../utils/types';
+import { KnownResult, Result, UnknownResult } from '../../utils/result';
+import { ICompilationSettings } from '../../utils/types';
+import { FileWithHash, TCVerdicts } from '../../utils/types.backend';
 
-export type LangCompileResult = Result<{ outputPath: string; hash: string }>;
+export interface LangCompileData {
+    outputPath: string;
+    hash?: string;
+}
+export type LangCompileResult = Result<LangCompileData>;
 const logger = new Logger('lang');
 
 export interface CompileAdditionalData {
     canUseWrapper: boolean;
-    compilationSettings?: {
-        compiler?: string;
-        compilerArgs?: string;
-        runner?: string;
-        runnerArgs?: string;
-    };
+    compilationSettings?: ICompilationSettings;
     debug?: boolean;
 }
 export const DefaultCompileAdditionalData: CompileAdditionalData = {
     canUseWrapper: false,
 };
+
 export class Lang {
     protected static executor: ProcessExecutor = new ProcessExecutor();
 
@@ -49,8 +52,11 @@ export class Lang {
         outputPath: string,
         additionalHash: string,
         forceCompile: boolean | null,
-    ) {
-        logger.trace('checkHash', {
+    ): Promise<{
+        skip: boolean;
+        hash: string;
+    }> {
+        logger.trace('Checking hash for file', src, {
             src,
             outputPath,
             additionalHash,
@@ -85,23 +91,57 @@ export class Lang {
         });
         return { skip: false, hash };
     }
+
     public readonly name: string = 'generic';
     public readonly extensions: string[] = [];
     public readonly enableRunner: boolean = false;
+
     public async compile(
         src: FileWithHash,
         ac: AbortController,
         forceCompile: boolean | null,
         additionalData: CompileAdditionalData = DefaultCompileAdditionalData,
     ): Promise<LangCompileResult> {
+        // Save the file if it's opened in an editor
         const editor = window.visibleTextEditors.find(
             (editor) => editor.document.fileName === src.path,
         );
         if (editor) {
             await editor.document.save();
+            // We have to wait until the document is actually saved to disk.
+            // Although `editor.document.save()` returns a promise, it seems that
+            // the promise resolves before the file is fully written.
+            // See https://github.com/langningchen/cph-ng/issues/60
             await waitUntil(() => !editor.document.isDirty);
         }
-        return this._compile(src, ac, forceCompile, additionalData);
+
+        // Clear previous compilation IO
+        CompilationIo.clear();
+
+        try {
+            const langCompileResult = await this._compile(
+                src,
+                ac,
+                forceCompile,
+                additionalData,
+            );
+
+            // Check if the output file exists
+            if (
+                langCompileResult instanceof UnknownResult &&
+                !(await exists(langCompileResult.data.outputPath))
+            ) {
+                return new KnownResult(
+                    TCVerdicts.CE,
+                    l10n.t('Compilation failed'),
+                );
+            }
+            return langCompileResult;
+        } catch (e) {
+            logger.error('Compilation failed', e);
+            CompilationIo.append((e as Error).message);
+            return new KnownResult(TCVerdicts.CE, l10n.t('Compilation failed'));
+        }
     }
     protected async _compile(
         _src: FileWithHash,
@@ -111,9 +151,46 @@ export class Lang {
     ): Promise<LangCompileResult> {
         throw new Error('Compile method not implemented.');
     }
-    public async runCommand(
+    protected async _executeCompiler(
+        cmd: string[],
+        ac: AbortController,
+    ): Promise<Result<undefined>> {
+        const result = await ProcessExecutor.execute({
+            cmd,
+            ac,
+            timeout: Settings.compilation.timeout,
+        });
+        if (result instanceof Error) {
+            return new KnownResult(TCVerdicts.SE, result.message);
+        }
+        if (result.abortReason === AbortReason.UserAbort) {
+            return new KnownResult(
+                TCVerdicts.RJ,
+                l10n.t('Compilation aborted by user'),
+            );
+        }
+        if (result.abortReason === AbortReason.Timeout) {
+            return new KnownResult(
+                TCVerdicts.CE,
+                l10n.t('Compilation timed out'),
+            );
+        }
+        CompilationIo.append(await readFile(result.stdoutPath, 'utf-8'));
+        CompilationIo.append(await readFile(result.stderrPath, 'utf-8'));
+        if (result.codeOrSignal) {
+            return new KnownResult(
+                TCVerdicts.CE,
+                l10n.t('Compiler exited with code {code}', {
+                    code: result.codeOrSignal,
+                }),
+            );
+        }
+        return new UnknownResult(undefined);
+    }
+
+    public async getRunCommand(
         _target: string,
-        _compilationSettings?: CompileAdditionalData['compilationSettings'],
+        _compilationSettings?: ICompilationSettings,
     ): Promise<string[]> {
         throw new Error('Run method not implemented.');
     }